requirements.txt
langgraph>=0.2
langchain-core>=0.2
requests>=2.31
python-dotenv>=1.0
gitpython>=3.1
pytest>=8.0
pytest-cov>=5.0
